datatype: suppress warnings on casting ptr to enum
Intel compiler warns:
    warning #810: conversion from void * to yaksa_type_t={enum
<unnamed>} may lose significant bits

+++ src/mpi/datatype/typerep/src/typerep_flatten.c
@@ -35,7 +35,7 @@ int MPIR_Typerep_flatten_size(MPIR_Datatype * datatype_ptr, int *flattened_type_
 
 #if (MPICH_DATATYPE_ENGINE == MPICH_DATATYPE_ENGINE_YAKSA)
     uintptr_t size;
-    int rc = yaksa_flatten_size((yaksa_type_t) datatype_ptr->typerep.handle, &size);
+    int rc = yaksa_flatten_size((yaksa_type_t) (intptr_t) datatype_ptr->typerep.handle, &size);
     MPIR_ERR_CHKANDJUMP(rc, mpi_errno, MPI_ERR_INTERN, "**yaksa");
     assert(size <= INT_MAX);
     flattened_loop_size = (int) size;
@@ -75,7 +75,8 @@ int MPIR_Typerep_flatten(MPIR_Datatype * datatype_ptr, void *flattened_type)
     flatten_hdr->num_contig_blocks = datatype_ptr->typerep.num_contig_blocks;
 
 #if (MPICH_DATATYPE_ENGINE == MPICH_DATATYPE_ENGINE_YAKSA)
-    int rc = yaksa_flatten((yaksa_type_t) datatype_ptr->typerep.handle, flattened_typerep);
+    int rc = yaksa_flatten((yaksa_type_t) (intptr_t) datatype_ptr->typerep.handle,
+                           flattened_typerep);
     MPIR_ERR_CHKANDJUMP(rc, mpi_errno, MPI_ERR_INTERN, "**yaksa");
 #else
     mpi_errno = MPIR_Dataloop_flatten(datatype_ptr, flattened_typerep);
